Fixed a bunch of problems. Changed from double to float. Changed the custom export format to JSON. Added visualization of the network via Display.html

+++ src/main/java/com/jeromewagener/util/Evaluator.java
@@ -9,10 +9,10 @@ import java.io.IOException;
 public class Evaluator {
     private ImageCompressor imageCompressor = new ImageCompressor(false);
     private boolean evaluatedAsCorrect;
-    private double certainty;
+    private float certainty;
 
     public void evaluate(String fileName, int realValue, Network network) throws IOException {
-        double[] inputVector = imageCompressor.compress(fileName);
+        float[] inputVector = imageCompressor.compress(fileName);
 
         if (network != null) {
             Network.Output networkOutput = network.calculate(inputVector);
